backend/requirements.txt
fastapi
uvicorn[standard]
torch==2.9.1
torchvision==0.24.1
opencv-python-headless
moviepy
pillow
lama-cleaner
numpy
python-multipart